// frontend/types.ts
export interface MicroCycle {
  id: string;
  name: string;
  startDate: string;
  endDate: string;
  focus: string[];
  volume: number;
  intensity: 'Low' | 'Medium' | 'High';
}

export interface MesoCycle {
  id: string;
  name: string;
  startDate?: string;
  endDate?: string;
  micros: MicroCycle[];
  isExpanded?: boolean;
}

export interface MacroCycle {
  id: string;
  name: string;
  season: string;
  startDate?: string;
  endDate?: string;
  model?: string;
  architecture?: string;
  mesos: MesoCycle[];
  isExpanded?: boolean;
}